Fix: Windows std::stoul long is smaller than uint64

+++ src/Nodes/DataProvider/IMU/FileReader/VectorNavFile.cpp
@@ -430,15 +430,15 @@ std::shared_ptr<NAV::NodeData> NAV::VectorNavFile::pollData(bool peek)
 
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_TIMESTARTUP)
                 {
-                    obs->timeOutputs->timeStartup = std::stoul(extractCell());
+                    obs->timeOutputs->timeStartup = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_TIMEGPS)
                 {
-                    obs->timeOutputs->timeGps = std::stoul(extractCell());
+                    obs->timeOutputs->timeGps = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_GPSTOW)
                 {
-                    obs->timeOutputs->gpsTow = std::stoul(extractCell());
+                    obs->timeOutputs->gpsTow = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_GPSWEEK)
                 {
@@ -446,11 +446,11 @@ std::shared_ptr<NAV::NodeData> NAV::VectorNavFile::pollData(bool peek)
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_TIMESYNCIN)
                 {
-                    obs->timeOutputs->timeSyncIn = std::stoul(extractCell());
+                    obs->timeOutputs->timeSyncIn = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_TIMEGPSPPS)
                 {
-                    obs->timeOutputs->timePPS = std::stoul(extractCell());
+                    obs->timeOutputs->timePPS = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.timeField & vn::protocol::uart::TimeGroup::TIMEGROUP_TIMEUTC)
                 {
@@ -578,7 +578,7 @@ std::shared_ptr<NAV::NodeData> NAV::VectorNavFile::pollData(bool peek)
                 }
                 if (binaryOutputRegister.gpsField & vn::protocol::uart::GpsGroup::GPSGROUP_TOW)
                 {
-                    obs->gnss1Outputs->tow = std::stoul(extractCell());
+                    obs->gnss1Outputs->tow = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.gpsField & vn::protocol::uart::GpsGroup::GPSGROUP_WEEK)
                 {
@@ -876,7 +876,7 @@ std::shared_ptr<NAV::NodeData> NAV::VectorNavFile::pollData(bool peek)
                 }
                 if (binaryOutputRegister.gps2Field & vn::protocol::uart::GpsGroup::GPSGROUP_TOW)
                 {
-                    obs->gnss2Outputs->tow = std::stoul(extractCell());
+                    obs->gnss2Outputs->tow = static_cast<uint64_t>(std::stoull(extractCell()));
                 }
                 if (binaryOutputRegister.gps2Field & vn::protocol::uart::GpsGroup::GPSGROUP_WEEK)
                 {
